pyproject.toml
[project]
name = "llm-tracekit"
version = "1.7.0"
description = "Collect and send GenAI calls as spans for development and production observability."
readme = "README.md"
authors = [
    { name="Coralogix Ltd.", email="info@coralogix.com" },
]
license = "Apache-2.0"
license-files = ["LICENSE.md"]
requires-python = ">3.9,<=3.13"
dependencies = [
    "opentelemetry-api>=1.32.1",
    "opentelemetry-exporter-otlp>=1.32.1",
    "opentelemetry-instrumentation>=0.53b1",
    "opentelemetry-sdk>=1.32.1",
    "opentelemetry-semantic-conventions>=0.53b1",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Framework :: OpenTelemetry :: Instrumentations",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

[project.optional-dependencies]
openai = [
    "openai>=1.26.0",
]
openai_agents = [
    "openai>=1.26.0",
    "openai-agents>=0.1.0; python_version >= '3.10'",
]
bedrock = [
    "boto3>=1.36.0",
]
gemini = [
    "google-genai>=1.0.0,<2",
]
litellm = [
    "litellm[proxy]>=1.65.4", # LiteLLM callbacks require proxy-specific dependencies
]
langchain = [
    "langchain>=1.0.0; python_version >= '3.10'",
]
langchain-openai = [
    "langchain>=1.0.0; python_version >= '3.10'",
    "langchain-openai>=1.0.0; python_version >= '3.10'",
]
langchain-aws = [
    "langchain>=1.0.0; python_version >= '3.10'",
    "langchain-aws>=1.0.0; python_version >= '3.10'",
]
langchain-all = [
    "langchain>=1.0.0; python_version >= '3.10'",
    "langchain-openai>=1.0.0; python_version >= '3.10'",
    "langchain-aws>=1.0.0; python_version >= '3.10'",
]

guardrails = [
    "pydantic>=2.10.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.27.0",
    "tenacity>=8.0.0",
]
all = [
    "openai>=1.26.0",
    "openai-agents>=0.1.0; python_version >= '3.10'",
    "boto3>=1.36.0",
    "google-genai>=1.0.0,<2",
    "langchain>=1.0.0; python_version >= '3.10'",
    "langchain-openai>=1.0.0; python_version >= '3.10'",
    "langchain-aws>=1.0.0; python_version >= '3.10'",
    "litellm[proxy]>=1.65.4",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.27.0",
    "tenacity>=8.0.0",
]


[project.urls]
homepage = "https://coralogix.com"
repository = "https://github.com/coralogix/llm-tracekit.git"


[project.entry-points.opentelemetry_instrumentor]
openai = "llm_tracekit:OpenAIInstrumentor"
openai_agents = "llm_tracekit:OpenAIAgentsInstrumentor"
boto3 = "llm_tracekit:BedrockInstrumentor"
gemini = "llm_tracekit:GeminiInstrumentor"
litellm = "llm_tracekit:LiteLLMInstrumentor"
langchain = "llm_tracekit:LangChainInstrumentor"

[dependency-groups]
dev = [
    "assertpy>=1.1",
    "botocore-stubs>=1.38.4",
    "mypy>=1.14.1",
    "pytest>=8.3.4",
    "pytest-asyncio>=1.0.0",
    "pytest-vcr>=1.0.2",
    "pyyaml>=6.0.2",
    "ruff>=0.9.7",
    "types-boto3>=1.38.4",
    "brotli>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
packages = ["llm_tracekit", "guardrails", "core"]

[tool.hatch.build.targets.wheel]
packages = ["llm_tracekit", "guardrails", "core"]

[[tool.mypy.overrides]]
module = ["wrapt", "opentelemetry.exporter.otlp.proto.grpc.trace_exporter"]
ignore_missing_imports = true

[tool.hatch.build.targets.sdist]
only-include = [
    "src",
    "tests",
]

[tool.uv]
override-dependencies = [
    "boto3>=1.40.19",
]
